+++ Fluency.info.php
@@ -6,7 +6,7 @@ namespace ProcessWire;
 
 $info = [
   'title' => 'Fluency',
-  'version' => '107',
+  'version' => '108',
   'href' => 'https://processwire.com/talk/topic/24567-fluency-integrated-deepl-powered-content-translation',
   'icon' => 'language',
   'summary' => 'The complete translation enhancement suite for ProcessWire.',
@@ -17,7 +17,7 @@ $info = [
     'LanguageSupport',
     'LanguageTabs',
     'ProcessWire>=300',
-    'PHP>=8.1'
+    'PHP>=8.1',
   ],
   'permission' => 'fluency-translate',
   'permissions' => [