+++ app/DataTransferObjects/FluencyConfigData.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Fluency\DataTransferObjects;
 
+use stdClass;
+
 /**
  * Provides an immunable object representing the current state of Fluency's configuration
  */
@@ -11,18 +13,18 @@ namespace Fluency\DataTransferObjects;
 final class FluencyConfigData extends FluencyDTO {
 
   /**
-   * @param bool    $translation_api_ready         Whether that translation engine API is ready
-   * @param bool    $translation_cache_enabled     Translation cache will be used
-   * @param ?object $selected_engine               The engine selected in Fluency config
-   * @param ?object $engine                        Current values for the translation engine config
-   *                                               variables
+   * @param bool                         $translation_api_ready     Whether that translation engine API is ready
+   * @param bool                         $translation_cache_enabled Translation cache will be used
+   * @param FluencyEngineSelectData|null $selected_engine           The engine selected in Fluency config
+   * @param stdClass|null $engine                                   Current values for the translation
+   *                                                                engine config variables
    * @param ?string $inputfield_translation_action translate_each_language or translate_to_all_languages
    */
   private function __construct(
     public readonly bool $translation_api_ready,
     public readonly bool $translation_cache_enabled,
-    public readonly ?object $selected_engine,
-    public readonly ?object $engine,
+    public readonly ?FluencyEngineSelectData $selected_engine,
+    public readonly ?stdClass $engine,
     public readonly ?string $inputfield_translation_action,
   ) {}
 
@@ -30,9 +32,6 @@ final class FluencyConfigData extends FluencyDTO {
    * {@inheritdoc}
    */
   public static function fromArray(array $config): self {
-    $config['selected_engine'] ??= null;
-    $config['selected_engine'] && $config['selected_engine'] = unserialize($config['selected_engine']);
-
     return new self(
       translation_api_ready: (bool) $config['translation_api_ready'],
       translation_cache_enabled: (bool) $config['translation_cache_enabled'],